# openapi.yaml
openapi: 3.0.3
info:
  title: UseZoracle Telegram Bot API
  description: API for managing cryptocurrency accounts, checking balances, and performing transactions using the Coinbase CDP SDK
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://usezoracle-telegrambot.up.railway.app
    description: Production server
...
paths:
...
  /api/snipe:
    post:
      summary: Snipe a token - execute a trade as fast as possible
      description: Executes a fast token purchase using secure CDP account management (no private keys required)
      tags:
        - Snipe
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST "http://localhost:3000/api/snipe" \
              -H "Content-Type: application/json" \
              -d '{
                "accountName": "my-snipe-account",
                "tokenAddress": "0x4200000000000000000000000000000000000006",
                "amount": "0.1",
                "slippage": 0.05
              }'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountName
                - tokenAddress
                - amount
              properties:
                accountName:
                  type: string
                  description: The CDP account name to use for the snipe
                  pattern: "^[a-zA-Z0-9-]+$"
                  example: "my-snipe-account"
                tokenAddress:
                  type: string
                  description: The token contract address to snipe
                  pattern: "^0x[a-fA-F0-9]{40}$"
                  example: "0x4200000000000000000000000000000000000006"
                amount:
                  type: string
                  description: Amount of ETH to spend on the token
                  example: "0.1"
                slippage:
                  type: number
                  description: Slippage tolerance (default 0.05 for 5%)
                  minimum: 0
                  maximum: 1
                  example: 0.05
      responses:
        "200":
          description: Snipe transaction executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnipeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/tokens/test/{contractAddress}:
    get:
      summary: Test token metadata
      description: Test token metadata fetching for any token address
      tags:
        - Tokens
      parameters:
        - in: path
          name: contractAddress
          required: true
          schema:
            type: string
          description: Token contract address to test
          example: "0x4200000000000000000000000000000000000006"
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/tokens/test/0x4200000000000000000000000000000000000006"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/tokens/test/0x4200000000000000000000000000000000000006')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Token metadata test completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Human-readable message about the operation
        data:
          type: object
          description: Response data
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp when the error occurred
...
    SnipeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Successfully sniped token 0x4200000000000000000000000000000000000006"
        data:
          type: object
          properties:
            tokenAddress:
              type: string
              description: Address of the token that was sniped
              example: "0x4200000000000000000000000000000000000006"
            amount:
              type: string
              description: Amount of ETH spent on the token
              example: "0.1"
            transactionHash:
              type: string
              description: Hash of the snipe transaction
              example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
            timestamp:
              type: number
              description: Timestamp when the snipe was executed
              example: 1703123456
...